Add flatpak-external-data-checker support
Add x-checker-data property for f-e-d-c.

Anitya checker was chosen as currently the git checker doesn't support
timestamp, so flathubbot will open a new PR on every run if the previous
one wasn't merged.
We could stay on git source type and use the JSON checker, but Github
has a 60 API requests hourly limit for unauthenticated requests, so it's
not fun to work with locally.