Update to v0.10.0
First time updating (manually) with f-e-d-c, next update PRs will be
opened automatically by flathubbot.
I chose not to have flathubbot auto-merge PRs with a successful CI build run.

I had to edit the `Exec` entry in the desktop file as it's hardcoding
`/usr/bin` path.
It seems to work correctly with `desktop-file-edit` even though the
desktop file has multiple groups, it only change the key value in the
`Desktop Entry` group.

A cleanup property was added to remove unneeded man page and shell
completions, as this is the convention.