Allow access to xdg-pictures
336226bf506a9b0cc3282cbc662d96993a08d31f revoked host access because it
was too broad. The app needs access to one directory to make it the
permanent save location. Since that "choosing" goes through the
portal-ed filechooser, a mounted path from the portal is returned which
is lost on restarts without filesystem access.

So allow access to the most common location to save screenshots